refactor: move common/entity.ts to building-blocks/template-model/pieces #565
affects: @haulmont/jmix-front-generator
