+++ packages/jmix-front-generator/src/building-blocks/stages/template-model/pieces/entity.test.ts
@@ -2,15 +2,15 @@ import {
   getDisplayedAttributes,
   isStringIdEntity,
   ScreenType
-} from "../../common/entity";
+} from "./entity";
 import {expect} from "chai";
 import {fail} from "assert";
 
-const dtViewProperties = require('../fixtures/view-properties--datatypes-test-entity.json');
-const o2oViewProperties = require('../fixtures/view-properties--association-o2o.json');
-const dtEntityModel = require('../fixtures/entity-model--datatypes-test-entity.json');
-const o2oEntityModel = require('../fixtures/entity-model--association-o2o.json');
-const projectModel = require('../fixtures/project-model--scr.json');
+const dtViewProperties = require( '../../../../test/fixtures/view-properties--datatypes-test-entity.json');
+const o2oViewProperties = require('../../../../test/fixtures/view-properties--association-o2o.json');
+const dtEntityModel = require('../../../../test/fixtures/entity-model--datatypes-test-entity.json');
+const o2oEntityModel = require('../../../../test/fixtures/entity-model--association-o2o.json');
+const projectModel = require('../../../../test/fixtures/project-model--scr.json');
 
 describe('getDisplayedAttributes()', () => {
   let dtBrowserDisplayedFields: string[];
