+++ packages/jmix-front-generator/src/generators/react-typescript/common/base-entity-screen-generator/index.ts
@@ -2,7 +2,7 @@ import Base from "yeoman-generator";
 import {BaseEntityScreenAnswers} from './params';
 import {Entity, EntityAttribute, ProjectModel} from '../../../../common/model/cuba-model';
 import {collectAttributesFromHierarchy, findEntity} from '../../../../common/model/cuba-model-utils';
-import {isStringIdEntity} from '../../../../common/entity';
+import {isStringIdEntity} from '../../../../building-blocks/stages/template-model/pieces/entity'
 import {fromStudioProperties} from '../../../../common/questions';
 import {idAttrNameQuestions} from '../questions';
 import {EntityTemplateModel} from '../template-model';
