+++ packages/jmix-front-generator/src/generators/react-typescript/entity-cards-grid/answers.ts
@@ -4,7 +4,7 @@ import {StudioTemplateProperty, StudioTemplatePropertyType} from "../../../commo
 import {CommonGenerationOptions} from "../../../common/cli-options";
 import {askQuestions} from "../../../building-blocks/stages/answers/defaultGetAnswersFromPrompt";
 import { askStringIdQuestions, StringIdAnswers } from "../../../building-blocks/stages/answers/pieces/stringId";
-import { isStringIdEntity } from "../../../common/entity";
+import { isStringIdEntity } from "../../../building-blocks/stages/template-model/pieces/entity";
 import {
   entityQuestion,
   EntityAnswer,
