+++ packages/jmix-front-generator/src/generators/react-typescript/entity-cards/answers.ts
@@ -4,7 +4,7 @@ import {StudioTemplateProperty} from "../../../common/studio/studio-model";
 import {CommonGenerationOptions} from "../../../common/cli-options";
 import {askQuestions} from "../../../building-blocks/stages/answers/defaultGetAnswersFromPrompt";
 import {askStringIdQuestions, StringIdAnswers} from "../../../building-blocks/stages/answers/pieces/stringId";
-import { isStringIdEntity } from "../../../common/entity";
+import { isStringIdEntity } from "../../../building-blocks/stages/template-model/pieces/entity";
 import {
   entityQuestion,
   EntityAnswer,
